requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.5
djangorestframework==3.12.2
freeze==3.0
future==0.18.2
html5lib==1.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.4
pytz==2020.5
reportlab==3.5.59
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
webencodings==0.5.1
xhtml2pdf==0.2.5
